requirements.txt
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
rich==13.9.2
